Clean-up_first - first attempt

+++ src/main/java/com/kodilla/ecommercee/domain/UserDto.java
@@ -6,9 +6,8 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class UserDto {
-    private int id;
+    private Long id;
     private String username;
     private boolean blocked;
     private String password;
-
 }
